+++ components/studio/datasets/datasheet_questions.py
@@ -1,4 +1,0 @@
-q = [
-    "<b>What do the instances that comprise the dataset represent (e.g.,documents, photos, people, countries)?</b> Are there multiple types of instances (e.g., movies, users, and ratings; people and interactions between them; nodes and edges)? Please provide a description.",
-    "<b>Who created the dataset (e.g., which team, research group) and onbehalf of which entity (e.g., company, institution, organization)?</b>"
-]