+++ docs/releasenotes.md
@@ -1,6 +1,16 @@
 # Release Notes
 
-## v0.4.0 (upcoming release)
+## v0.5.0 (upcoming release)
+
+### New functionality
+- Studio Administration module to monitor and manage resources as an alternative to Django Admin
+- Brand new STACKn UI
+
+### Bug fixes
+
+### Other
+
+## v0.4.0
 
 ### New functionality